Mann-Kendall Test of Prewhitened Time Series Data in Presence of Serial Correlation Using the von Storch (1995) Approach

When time series data are not random and influenced by autocorrelation, prewhitening the time series prior to application of trend test is suggested.

pwmk(x)

Arguments

x
- Time series data vector

Value

Z-Value - Z statistic after prewhitening

Sen's Slope - Sen's slope for prewhitened series

old. Sen's Slope - Sen's slope for original data series (x)

P-value - P-value after prewhitening

S - Mann-Kendall S statistic

Var(s) - Variance of S

Tau - Mann-Kendall's Tau

Details

The lag-1 serial correlation coefficient is used for prewhitening.

References

Kendall, M. (1975). Rank Correlation Methods. Griffin, London, 202 pp.

Kulkarni, A. and H. von Storch. 1995. Monte carlo experiments on the effects of serial correlation on the MannKendall test of trends. Meteorologische Zeitschrift N.F, 4(2): 82-85.

Mann, H. B. (1945). Nonparametric Tests Against Trend. Econometrica, 13(3): 245-259.

Salas, J.D. (1980). Applied modeling of hydrologic times series. Water Resources Publication, 484 pp.

von Storch, V. H. (1995). Misuses of statistical analysis in climate research, In: Analysis of Climate Variability: Applications of Statistical Techniques, ed. von H. V. Storch and A. Navarra A. Springer-Verlag, Berlin: 11-26.

Yue, S. and Wang, C. Y. (2002). Applicability of prewhitening to eliminate the influence of serial correlation on the Mann-Kendall test. Water Resources Research, 38(6), <doi:10.1029/2001WR000861>

Examples

x<-c(Nile)
pwmk(x)

#>          Z-Value      Sen's Slope old. Sen's Slope          P-value 
#>      -2.55152628      -1.39262731      -2.60000000       0.01072522 
#>                S           Var(S)              Tau 
#>    -845.00000000  109417.00000000      -0.17419089